feat: [CCM-7535]: Refactor And Fix Old Tests
